moves test helpers to src
