+++ config/services.yaml
@@ -29,9 +29,6 @@ services:
         resource: '../src/Controller'
         tags: ['controller.service_arguments']
 
-    App\Tests\Helpers\:
-        resource: '../tests/Helpers'
-
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
 
@@ -212,4 +209,4 @@ services:
 
     App\DataFixtures\BehatFixtures:
         arguments:
-            $fixtureHelper: '@App\Tests\Helpers\FixtureTestHelper'
+            $fixtureHelper: '@App\TestHelpers\FixtureTestHelper'
