+++ src/TestHelpers/UserTestHelper.php
@@ -1,7 +1,7 @@
 <?php declare(strict_types=1);
 
 
-namespace App\Tests\Helpers;
+namespace App\TestHelpers;
 
 use App\Entity\User;
 use App\Tests\ApiWebTestCase;
