+++ tests/Service/OrderServiceTest.php
@@ -8,8 +8,8 @@ use App\exceptions\WrongCaseNumberException;
 use App\Service\DocumentReaderService;
 use App\Service\OrderService;
 use App\Service\SiriusService;
-use App\Tests\Helpers\FileTestHelper;
-use App\Tests\Helpers\OrderTestHelper;
+use App\TestHelpers\FileTestHelper;
+use App\TestHelpers\OrderTestHelper;
 use Doctrine\ORM\EntityManager;
 use Exception;
 use Prophecy\Argument;
